+++ packages/terra-form-checkbox/CHANGELOG.md
@@ -3,6 +3,9 @@ ChangeLog
 
 Unreleased
 ----------
+
+3.22.0 - (August 14, 2019)
+------------------
 ### Changed
 * Components updated to use `FormattedMessage` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts