+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+
+5.26.0 - (August 14, 2019)
+------------------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
+
 ### Added
 * Added Combobox and ComboboxField exports and doc site entries
 * Added SingleSelect and SingleSelectField exports and doc site entries