+++ packages/terra-search-field/CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Unreleased
 ----------
+
+3.24.0 - (August 14, 2019)
+------------------
+### Removed
+* Removed explicit calls to set `caret-color` to `transparent` in WDIO tests
+
 ### Changed
 * Updated search-field example code to show more usage.
 * Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.