+++ packages/terra-status-view/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+4.5.0 - (August 14, 2019)
+------------------
 ### Changed
 * Update jest test snapshot
 * updated package.json test scripts